backend/requirements.txt
fastapi
uvicorn
langchain
pinecone-client
openai
pymupdf
docx2txt
pytest
